requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
cloudinary==1.25.0
dj-database-url==0.5.0
Django==3.2.3
django-cors-headers==3.7.0
django-crispy-forms==1.9.1
django-filter==2.2.0
django-google-maps==0.12.1
django-heroku==0.3.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
docutils==0.16
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
Markdown==3.2.2
MarkupSafe==1.1.1
Pillow==7.1.1
psycopg2==2.8.6
pyftdi==0.52.0
Pygments==2.6.1
PyJWT==1.7.1
pyserial==3.5
pytz==2019.3
pyusb==1.1.0
requests==2.23.0
six==1.16.0
sqlparse==0.3.1
typing-extensions==3.10.0.0
urllib3==1.25.9
Werkzeug==1.0.1
whitenoise==5.2.0
zipp==3.1.0
